Remove touch target gaps between ActionBar items
