+++ .changeset/popular-timers-shop.md
@@ -1,0 +1,5 @@
+---
+'@primer/view-components': minor
+---
+
+Remove touch target gaps between ActionBar items
