+++ test/components/alpha/action_bar_test.rb
@@ -7,6 +7,43 @@ module Primer
     class ActionBarTest < Minitest::Test
       include Primer::ComponentTestHelpers
 
+      def test_renders
+        render_inline(Primer::Alpha::ActionBar.new) do |component|
+          component.with_item_icon_button(icon: :pencil, label: "Button 1")
+          component.with_item_icon_button(icon: :pencil, label: "Button 2")
+          component.with_item_icon_button(icon: :pencil, label: "Button 3")
+          component.with_item_icon_button(icon: :pencil, label: "Button 4")
+        end
+
+        assert_selector("action-bar") do
+          assert_selector("tool-tip", text: "Button 1")
+          assert_selector("tool-tip", text: "Button 2")
+          assert_selector("tool-tip", text: "Button 3")
+          assert_selector("tool-tip", text: "Button 4")
+          assert_selector("[data-target=\"action-bar.moreMenu\"]", visible: :hidden)
+        end
+      end
+
+      def test_size_small
+        render_inline(Primer::Alpha::ActionBar.new(size: :small)) do |component|
+          component.with_item_icon_button(icon: :pencil, label: "Button 1")
+          component.with_item_icon_button(icon: :pencil, label: "Button 2")
+          component.with_item_divider
+          component.with_item_icon_button(icon: :pencil, label: "Button 3")
+          component.with_item_icon_button(icon: :pencil, label: "Button 4")
+        end
+
+        assert_selector("[data-targets=\"action-bar.items\"] .Button--small", count: 4)
+      end
+
+      def test_item_merges_item_arguments
+        render_inline(Primer::Alpha::ActionBar.new(size: :small)) do |component|
+          component.with_item_icon_button(icon: :pencil, label: "Button 1", item_arguments: { classes: "foo", tag: :span })
+        end
+
+        assert_selector("span.foo.ActionBar-item")
+      end
+
       def test_renders_action_menu_items_with_type_button
         render_preview(:default)
